Create RoomBlocker, change Exit door, Add door sprite
RoomBlocker.cs is assigned to a wall which block exit room , ExitDoor.cs is asigned to exit door game object
