+++ Assets/Scripts/ExitDoor.cs
@@ -1,18 +1,30 @@
 using System;
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.SceneManagement;
+using Object = UnityEngine.Object;
 
 public class ExitDoor : MonoBehaviour
 {
+    [SerializeField] private Object nextScene;
     [SerializeField] private UI.UIMessage uiMessage;
-    private bool _isActive;
-    
-    private void Update()
+    private void OnTriggerEnter2D(Collider2D other)
     {
-        if (!_isActive)
+        if (other.CompareTag("Player"))
         {
-            if ((GameManager.EnemiesCount / 2) > GameManager.Kills) return;
-            uiMessage.UpdateMessage("You have Killed enough guards. Now leave the area.");
-            _isActive = true;
+            uiMessage.UpdateMessage("Press 'Enter' to exit.");
+        }
+    }
+
+    private void OnTriggerStay2D(Collider2D other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            if (Input.GetKeyDown(KeyCode.Return))
+            {
+                SceneManager.LoadScene(nextScene.name);
+            }
         }
     }
 }
