+++ Assets/Scripts/ExitDoor.cs.meta
@@ -1,3 +1,11 @@
 fileFormatVersion: 2
-guid: c1faf6f7d89340bdac654ab30e29c26e
-timeCreated: 1695430157
+guid: 1ace933d0ceb3a544a8bf51223a54927
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
