+++ UserSettings/EditorUserSettings.asset
@@ -12,9 +12,12 @@ EditorUserSettings:
       value: 5557055f55565e5d0f5e0920467b5c4445151e7c7b2a7666782c1c6bb3b66c39
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 5a0256075001590f0f5d5f2343210644144e1a282e7e74647a7e1f31e7b6673e
+      value: 00025203535408095c080e754170084413161d7a297b7f36797b4d35b5e2656a
       flags: 0
     RecentlyUsedSceneGuid-3:
+      value: 5a0256075001590f0f5d5f2343210644144e1a282e7e74647a7e1f31e7b6673e
+      flags: 0
+    RecentlyUsedSceneGuid-4:
       value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     vcSharedLogLevel:
